requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.28,<1.0
pydantic-settings>=2.6,<3.0
apscheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
